Implement QueryFragment for generated Enums.
So that we can just call selection_set on them rather than having to
special case enums in the QueryFragment derive.  Not sure this is the
best approach but it'll do for now.

Figured that just deriving serialize & deserialize for these types is
also not safe, so added a TODO around that.