Add support for union types via InlineFragments trait.
This adds an InlineFragments trait that can be used to describe a
GraphQL inline fragment.  This trait can currently be derived on an enum
in order to provide support for querying union types.  I've plans to
expand this to support interfaces in the future, but this isn't done
just yet.