Add InterfaceStruct to query_dsl.
This provides a stub struct to any interface definitions, that can be
used as a TypeLock on fields that provide these interfaces.  We still
don't support actually querying interfaces, this just means we can
output a query_dsl for a schema with interfaces without error.

Full support for interfaces will come in a future PR.