Simple type validation in QueryFragment derive.
When writing QueryFragments, it's really easy to get confused about
what GraphQL types are required, what are lists etc.  When you get this
wrong, the Rust compile errors are either not as helpful as they could
be _or_ plain confusing.

This updates the QueryFragment derive code to be a bit smarter here: it
will specifically check for the cases where there's a mismatch in list
types or optionalness.

I've also updated it to error when people provide generic types in their
structs.  Currently this will cause a syntax error in the derive output,
with an unhelpful message, so I'm just explicitly disallowing it for
now.  Can revisit fixing the underlying error at some point.