Added query_module attribute macro. (#15)
This adds a query_module attribute macro that can be applied to a module containing QueryFragments & InlineFragments.  This was intended as an optimisation so that each individual derive macro invocation didn't have to load the graphql schema for itself.  Turns out this was not really a bottleneck - with the github GQL test app I'm using the builds still take about 20 seconds.  Seems like the slow bit is rust parsing the query_dsl.

Despite this, query_modules do offer some advantages:

1. They allow us to consolidate some of the parameters that each of our derive macros take - instead of providing the query_dsl & schema_path on each struct you can do it once on the query_module macro.
2. They open up the possibility of dropping the `graphql_type` annotation as well - if we can figure out the types by walking the query struct tree then we could potentially figure this out for ourselves.  Haven't implemented this yet, but it seems like it might be a nice addition...