Revamped enum support (#19)
Prior to this change, we were generating enums as part of the generated query_dsl, and users were expected to use these generated enums.  I realised this was fairly inconsistent with the principle of users bringing their own structs, so this PR changes that.

Now, users are expected to provide their own enums, and `#[derive(cynic::Enum)]` on them to make them compatible with cynic & verify that all the variants are covered.  I've also updated cynic-querygen to do this automatically for any queried enums.