Always output all scalars & enums in querygen.
We're transitioning to users owning all the enums & scalars.  Prior to
this querygen was just outputting the scalars & enums that we needed to
service the query provided.

However, the query_dsl still requires all scalars & enums that are
required for input structs or arguments, so this updates querygen to
output all the enums & scalars into a `types` module, which the
query_dsl can input.  Not perfect, but can maybe revisit this when/if I
come up with a way to avoid.