Stop query_dsl from depending on user generated enums.
Prior to this change the query_dsl module depended on having access to
any enums that were required for input types.  This was a bit of a pain
- you had to define all the enums in the schema, even if you weren't
using them.

This updates the query_dsl to make anywhere that depended on concrete
enum types be generic, with constraits to our Enum trait.  I needed to
add a generic parameter to our Enum traits to enable us to type lock
them in query_dsl, and also update query_dsl to generate marker traits
for these types.

Also updated the Enum derive to generate these parameters.