No inputs in query dsl (#23)
Prior to this change the query_dsl module depended on having access to
any enums that were required for args or input types.  This was a bit of a pain
- you had to define all the enums in the schema, even if you weren't
using them.

The query_dsl also generated types for any input objects in the schema,
which went against the principal of letting users bring their own structs.

This updates the query_dsl to make anywhere that depended on concrete
enum or input types be generic, with constraits to our Enum trait and 
InputObject trait respectively.

I needed to add a generic parameter to our Enum traits to enable us to type lock
them in query_dsl, and also update query_dsl to generate marker traits
for these types.  Also updated the Enum derive to generate these parameters.

There's still some work to be done to make this properly usable, particularly for
input.  Decoding enums will work, but we're still using serde to Serialize things,
and it's unlikely to get it right.  Also need to update querygen to actually generate
types for any inputs we need.