Drop Serialize requirement for enums & input types. (#27)
Prior to this change, enums & input types needed to implement serde::Serialize, as it was
used by the Arguments struct to serialize things to JSON for sending to servers.

This didn't actually work without a lot of work though, as Serialize is not aware of the
transformations that cynic::Enum applies to enums, or any renames that an eventual
derive for cynic::InputType will apply, so you'd need to manually tell serde about all
these things.

This fixes that by adding a new cynic::SerializableArgument trait.  This is automatically
implemented for InputTypes using a blanket impl, and is derived for enums when deriving
cynic::Enum.  Wanted to do a blanket derive on cynic::Enum but couldn't for rust reasons.

Fixes #25