Rework query_dsl argument structs.
The previous implementation of arugment structs in the query DSL had
some issues:

1. Supplying arguments was pretty verbose, particularly where you wanted
   to supply required _and_ optional arguments or where you didn't want to
   supply args at all, but the field had some.
2. You also needed to import structs from their nested modules, which
   was also probably not great.
3. The optional argument structs tried to derive Default, but due to
   issues with the default derive this forced all argument types to also
   implement default (See #40)
4. Finally, there were some serious problems around enum arguments where
   we needed to proivde a generic parameter but were completely unable to
   as the argument wasn't being provided.  (See #36)

This updates the query DSL so that each field selector function
takes required arguments, and then returns a FieldSelectionBuilder.
This FieldSelectionBuilder will expose a function for each optional
argument, and a select function that will output the corresponding
SelectionSet for use in the query heirarchy.