Serialize arguments on query serialize.
Prior to this change we serialized arguments when building up the
selection set, and stored the results of that serialization in the
Argument struct.  This worked, but led to an awkward API where you build
the selection set, convert it to a QueryBody which could fail, and
_then_ convert it to JSON which could also fail.

This updates the Argument struct to store a `Box<SerializableArgument>`
instead.  This simplifies things: we can now serialize the Query
directly, and the Argument serialization will happen there, leading to a
single point of failure rather than multiple.