Make SerializableArguments require Send. (#56)
#### Why do we need this change?

I've been adding cynic to an async app using surf.  In order for the query to persist over an await, it 
needs to be `Send`.  This is needed, as we need to be able to decode after the `Send`.

#### What effects does this change have?

This updates various generic & dyn constraints to require `Send` when a `SerializableAgument` is
used.  This solves the above problem, though I'm not 100% happy with the approach.