Support hardcoded query values in cynic-querygen.
This fixes an issue where we were trying to lookup enums by the name of
one of their variants rather than by the name of the enum itself.

Also tidied up a lot of the stuff around the TypeIndex in querygen - now
uses raw schema types for a lot of things, and supports returning the
actual Field value rather than just the Type of the Field.

Fixes #33