Rework argument code for more lenient typing (#76)
#### Why do we need this change?

Currently, the `cynic_arguments` attribute is pretty strict in what it accepts.  

- If an argument is `String!` you need to explicitly provide a `String` (a `&str` isn't good enough). 
- If an argument is `String` you need to explicitly provide an `Option<String>` - none of `String`, '&str`, or `Option<&str>` is acceptable.

This leads to some noisy `cynic_argument` attributes.

This also causes otherwise backwards compatible GraphQL schema changes to cause compilation errors - if a field was required before and is now optional, you'll have to explicitly add that `Some` to your definition.

#### What effects does this change have?

This PR adds an `IntoArguments<T>` trait where `T` is the type of the argument.  This trait handles converting various arguments into their expected form.

I considered using some of the built in traits like `AsRef` or `Into` but they all seemed to come with limitations for what I wanted to do.

Fixes #48 

#### Note on Lifetimes

I spent some time trying to solve #46 at the same time - at one point `IntoArgument` had a `Output` associated type that I could use to control the type being serialized.  I used this to borrow `&str` and feed it into a new lifetime on `Argument`.  This new lifetime ended up bleeding up to the `Query` struct which made things a bit unergonomic.  See note on #46 for more details on that.  At the time of writing this alternative implementation can be found on the `into-arguments` branch, though it may not stay around forever.  This PR was a partial rebase of that original branch, with some re-work afterwards.