Comma separate argument constructor calls in query_dsl.
Tested the query_dsl code against the GitHub schema, but was getting
"unexpected token ::" errors from the query_dsl macro output.

Tracked it down to the `Argument::new` calls in the `FieldSelector`
output - wasn't separating them with commas so if any fields had > 1
required argument it would fail to compile.  Added a comma to fix