Add GitHub example (#96)
#### Why are we making this change?

The current example uses the StarWars API.  This is good, but doesn't have a
lot of GraphQL functionality - in particular InputObjects aren't used at all.
I'm in the middle of adding InputObject support so it'd be good to have an
example (and therefore a test) of InputObject support.

#### What effects does this change have?

It adds a GitHub example that queries the cynic PRs to the repo.  This both
documents the feature and allows us to test it by running the example.

Unfortunately the GitHub API is too big to compile a cynic example quickly so I
had to resort to feature flagging bits of the example.  This diminishes it's
usefulness as documentation so I should probably look into how to do this
better in the future.

This test currently does nothing as the feature flag I've added is only declared 
in #95.  Going to merge this and then merge it into #95 to test.