Add support for InputObjects (#95)
#### Why are we making this change?

GraphQL input objects were fairly badly supported up till now.  It was probably
possible to write them manually, but no derives were there and the process was
fairly undocumented.

#### What effects does this change have?

This PR contains a derive for InputObject that allows easy creation of
InputObjects.  It also adds support for this derive to querygen.

Fixes #29 
Fixes #30