Removed optimised query modules (#100)
#### Why are we making this change?

At one point the query_module attribute called into the other derive attrs
itself so as to avoid having to parse the GQL schema several times. This didn't
seem to have much performance impact, but did make error reporting 
significantly worse - all the errors in a module pointed to the attr, rather
than at least the struct they came from. So I put this impl behind a feature
flag.

Later on I was making some changes that added a new path to the optimised query
module feature in addition to an impl outwith it.  I'm not convinced the
feature is that useful, so I removed it rather than adding this new path to it.

#### What effects does this change have?

Strips out the code that powered the feature.  The original feature flag was
removed in #95.

Fixes #98