Add tests using graphql.jobs API.
The starwars schema is great, but it doesn't cover all the types of
queries you can make via GraphQL.  In this case I'm wanting to test
input objects, but it just has none.

So I've added some tests with the graphql.jobs API, which does have some
input objects in it.  Was using the GitHub API for this previously, but
since it requires an API key it's not quite as easy to use in tests.