Automated builds of querygen output for testing (#101)
#### Why are we making this change?

Up till now I've not been doing particularly thorough testing of the output 
from querygen.  The expectation has been that people can use querygen to get
close to what they want and then fix up all the problems themselves.  This has
been fine so far, but isn't a very good long term strategy.  I'd like querygen
to be a great way to learn how to write queries with cynic, and that means it
needs to get things right.

#### What effects does this change have?

This PR adds tests that build & run the output from querygen.  We take a query, 
run it through querygen, nd then use the `trybuild` crate to build it and run
it.`

It also adds tests using the GraphQL jobs API, as this API has input objects and
the star wars API doesn't.

Fixes #61