Fix object literal construction in querygen.
The code we were outputting was missing optional fields as we just
print out the fields that are present in the JSON given to us by the
query.  This change updates to also print out any optional fields that
are not in the literal provided to us.

In the case of required fields we don't do this, as the compiler will
catch these errors with a good message.