Move Argument serialization to argument construction
Storing a copy of `SerializableArguments` inside `Arguments` seems to be
causing all sorts of issues: we can't take references to arguments, a
lot of things need to be Clone, Argument types need to be Send.

I want to see if I can drastically simplify the API by just doing the
serialization at the point of `Argument::new`.  Feels a bit of an odd
place to do it but functionally it's basically the same, and if it
allows for a better API then cool.