Don't take ownership of SerializableArguments in Argument (#120)
#### Why are we making this change?

Prior to this change `Argument` contained a `Box<SerializableArgument>`.  This
had a few knock on effects on the API of cynic:

1. A `SerializableArgument` had to be `'static`, as `Argument` had no lifetime
   parameter.
2. A `SerializableArgument` had to be `Send` if I wanted cynic to be used
   across async/await boundaries (which I do).
3. As `Argument` couldn't take anything by reference, most things that impl
   `SerializableArgument` had to be `Clone` - `InputObjects`, `Enums`, etc.

It was raised in #88 that this isn't great generally.

#### What effects does this change have?

This reworks `Argument` to serialize it's `SerializableArgument` at the point
of creating the `Argument` and store the result of this inside `Argument`.
This allows us to pass in references and avoid taking ownership, which lets us
get rid of a lot of constraints, and a lot of un-neccesary `clones`.

I've also updated `IntoArgument` to support returning references instead of
always returning owned objects as the constraints on Argument::new were the
only thing preventing this.