Add `surf` feature (#123)
#### Why are we making this change?

Cynic is currently HTTP client agnostic - we expect users to write their own
HTTP client code.  It's not particularly hard, but it's also not trivial - you
need to send a request with the right format, decode the response as JSON and
then decode it again using your `cynic::Operation`.

It would be easier for newcomers, make documentation easier and generally
provide a better API if cynic optionally came with HTTP client support.

#### What effects does this change have?

This adds support for `surf` to cynic, behind a feature flag.  This support
comes in the form of an extension trait for `surf::RequestBuilder`.  Extending
this struct allows users to use the normal `surf` API for setting any headers
etc. and then defer to `cynic` for only the bits that are required.

Part of #64