Finish building out the GraphQLError type.
This type is now complete according to the GraphQL spec, and includes
proper data models for locations, path & the extensions field. All of
these fields are optional, however their shape is well defined when
included. The extensions field is defined as being an arbitrarily shaped
JSON object, and as such is modeled as an `Option<serde_json::Value>`.

Also added the feature `surf-wasm-client` feature which will enable the
`surf/wasm-client` passthrough feature so that the surf feature may be
used in the wasm32 web context.

Added a convenience constructor for the GraphQLError type.

closes #127