Finish building out the GraphQLError type. (#129)
#### Why are we making this change?

These changes are being made to finish up the remaining work on the
GraphQLError type. It is now complete, and allows for full transmission of
error data from the server side to the client side.

#### What effects does this change have?

This type is now complete according to the GraphQL spec, and includes proper
data models for locations, path & the extensions field. All of these fields are
optional, however their shape is well defined when included. The extensions
field is defined as being an arbitrarily shaped JSON object, and as such is
modeled as an `Option<serde_json::Value>`.

Also added the feature `surf-wasm-client` feature which will enable the
`surf/wasm-client` passthrough feature so that the surf feature may be used in
the wasm32 web context.

This change is fully backwards compatible for all GraphQL spec compliant
servers. All of the added fields are optional, as defined in the spec. Overall,
this change will allow for the full use of GraphQL's error handling mechanisms,
which my production app, and many others, rely upon heavily.

A sample of what this error response object looks like in action (this is an
error response from a Juniper GraphQL API too):

```
GraphQLResponse {
    data: None,
    errors: Some([GraphQLError {
        message: "Invalid verification code provided.",
        locations: Some([GraphQLErrorLocation { line: 2, column: 3 }]),
        path: Some([Field("federateLogin")]),
        extensions: Some(Object({"status": String("400")}))
    }])
}
```

Closes #127