Make http extension futures Send (#130)
#### Why are we making this change?

I just tested out the new SurfExt trait with some of my production code.  I was
having trouble getting it working, as the futures it returns were not Send.
This was mostly because `SerializeError` was not `Send + Sync`.

#### What effects does this change have?

This adds `Send + Sync` constraints to `SerializeError` which allows the
futures in `SurfExt` & `ReqwestExt` to be `Send`.

Possible that the API could have been reworked to do this without updating the
constraints on `SerializeError`, but I'm fairly sure having errors as Send +
Sync is good practice anyway.