Querygen query parsing rewrite (#141)
#### Why are we making this change?

When I initially wrote the query generator code I was in a rush.  I wanted 
something that worked good enough to get started, but didn't focus too hard on
handling every single case.

#### What effects does this change have?

This commit is a large overhaul of the query generating code, taking care to do
the correct thing in a lot more cases.  Amongst other things it:

1. Deduplicates generated types
2. Supports multiple queries, sharing structs between all the generated queries
   as appropriate.
3. Generates unique names for each struct it creates, even when faced with
   different structs targetting the same type.
4. Generates partial InputObjects when faced with literals with missing fields
   (previously it would generate all fields even when unused)
5. Correctly generates different argument structs if a single type with
   arguments is used in multiple queries.  Though the correct IntoArgument impl
   is not yet generated.
5. Pathes the way for supporting fragments in the GraphQL input, though this is
   not supported yet