Refactor generator output code. (#159)
#### Why are we making this change?

I'm working on fixing #157, which is going to involve putting some more logic
around the output of field names in the generator.  At the moment all the
output code happens in a single big function that interpolates strings and puts
them in a Vec of lines.  I threw the code together in a hurry at some point and
it could do with a tidy.

#### What effects does this change have?

This starts the process of tidying this code:

1. Adds a new output module for this code to live in, mostly moved from the
   query_parsing module of the same name.
2. Adds Display impls of most of the Output structs
3. Updates the output function to use std::fmt::Write & these Display impls
4. Provides an indent std::fmt::Write wrapper that handles indentation a bit
   better.
5. Refactors the ArgumentStruct code a little so the _actual_ name of the
   struct is contained within the output struct.