Make InputObject look scalars up in the correct place (#162)
#### Why are we making this change?

The InputObject derive outputs some type checking functions. As part of this
it needs to use parameter types in two ways. First it needs to use the field
type taken from the InputObject itself. Second it needs to use the typelock
type defined in the `query_dsl` to provide generic bounds.

The typelock needs to be taken from `query_dsl`, but the literal field type
does not.  I messed up writing this code though and the literal field type was
pulled from the query_dsl.

#### What effects does this change have?

Updates the InputObject derive to only pull the typelock from the query_dsl.

No test as I have a test locally that still doesn't work - will commit it once
I finish fixing issues.

Fixes #139