Make ValueType::to_literal context aware. (#163)
#### Why are we making this change?

The generator was previously unware of the context in which it was printling
literal values.  This is an issue, because the correct thing to do depends
heavily on context: required arguments usually need cloned, optional arguments
are safe to be referenced but not if they're inside an object literal etc.

#### What effects does this change have?

This updates the `TypedValue::to_literal` function with a concept of the
context the literal is being printed in, and does a much more thorough job of
printing variables correctly within these contexts.

There was no issue that tracked this specifically.

This fixes #140 as we now have a test that includes a scalar inside an
InputObject (and are deriving Clone on scalars). This test also covers the
fixes in #160 #161, #162.