Implement list & option coercion in generator. (#165)
#### Why are we making this change?

GraphQL has the concept of coercion for list inputs: if a type expects a list
you can provide a single element of that list and the GraphQL server should
treat it like a list with a single element.  Similarly, GraphQL doesn't have
explicit option types so any Option<T> field in GraphQL is happy to accept a
literal T.

For option coercion I've implemented the `IntoArguments` trait: any field that
expects an Option<T> is happy to accept a `T` because there's often an
`IntoArguments<Option<T>> for T`.  I'd have liked to extend this to lists, but
unfortunately rust starts running into type inference problems when I do this.

#### What effects does this change have?

Updates the generator to correctly wrap variables & list types in lists when
required and also to cover for a couple of other places where option coercion
doesn't quite work.  This should mean that users can provide queries that rely
on list coercion to the generator and it will still output working rust.

Fixes #106