Set defaults for rename_all (#167)
#### Why are we making this change?

The `InputObject` & `Enum` derives have a top-level `rename_all` attribute.
This exists because Rust naming conventions and GraphQL naming conventions tend
to differ: rust uses camel_case for fields and PascalCase for variants, whereas
GraphQL uses pascalCase for fields and SCREAMING_SNAKE_CASE for variants.

Prior to this change it was expected that users would provide this field
themselves.  However, this is inconsistent with `QueryFragment` derives where
you don't need to provide anything.

#### What effects does this change have?

Updates the rename_all attribute to default to the appropriate value for
standard GraphQL conventions, bringing the `InputObject` & `Enum` derives in
line with the `QueryFragment` derives.  If users want to override it (or
disable it) they can.

Fixes #99