Allow optional InputObject arguments to be passed by reference. (#187)
#### Why are we making this change?

`IntoArgument` provides argument conversion for cynic to allow a bit of
flexibility in the types that can be accepted for any given argument.  Prior to
this change there was no `IntoArgument` definition for optional references to
`InputObjects` - all `Option<InputObject>` arguments had to be passed as owned
objects.  This meant you needed to clone the entire InputObject.  This also
meant that the generator output didn't work, as it assumed references should
work.

#### What effects does this change have?

We had a macro for dealing with these conversions internally, but it defined a
trait on `Option<&T>` so was not suitable for use outside of the crate.  This
moves the definitions in that macro into generic defintions for _any_
`Option<T>`.  This seems to solve the problem, and importantly doesn't any
existing functionality.

Fixes #185