Disable GraphiQL fetches if no schema URL provided (#188)
#### Why are we making this change?

The generator has two ways of provided a schema: via a URL or by pasting
GraphQL SDL into a text box.  When pasting the SDL we have no server to run
queries against, but GraphiQL tries to query `/` anyway.  On production this
leads to a netlify 404 appearing in the results pane, and periodic failed
fetches in the browsers dev tools as GraphiQL tries to run the introspection
query.

#### What effects does this change have?

Updates our fetcher to not run fetches if we have no schema, fixing this
behaviour.

Fixes #184