Validate for empty composites (#212)
#### Why are we making this change?

If you select a GraphQL object field (i.e. a composite) but don't select any                                                                                                                                                             fields within that object, that is not valid GraphQL.  It is a syntactically
valid query, but not semantically valid.  Cynic does not really check for this
at the moment, and it's quite easy to build a query like this with the GraphiQL
explorer in the generator.  The generator editor correctly labels it as
invalid, but then code is generated that will compile but not output a valid
query.

#### What effects does this change have?

This two of the places where this could be an issue:

1. The generator will no longer generate types for such queries.
2. If users provide an empty QueryFragmeng that will error.