Add TypeLock to `cynic::Scalar` (#218)
#### Why are we making this change?

As outlined in #182 - cynic currently requires all scalars in a schema to have
concrete types that are available for the query_dsl.  This means that you need
to explicitly define all scalars, even if you're not using them.  This also
means that new scalars added upstream are a breaking change for cynic clients,
when they normally would not be for GraphQL clients.

#### What effects does this change have?

This switches scalars to use a similar technique to `Enum`s and `InputObjects`.

The `Scalar` trait is now `Scalar<TypeLock>` where `TypeLock` should be some
type defined in the `query_dsl`.  Methods that previously took any scalar now
take a `Scalar<TypeLock>`.  This also plugs a hole in the previous scheme where
any Scalar could be provided for any particular scalar field.

This has had some fairly large knock on effects in the code, which will be
detailed in the CHANGELOG. 

Fixes #51 
Fixes #119
Fixes #182