Rename query_dsl! -> use_schema! (#232)
#### Why are we making this change?

Cynic has long had the idea of a `query_dsl`.  This name originated when I was intending
users to directly use the DSL that it output to generate queries.  That's no longer really my
expectation, and the name query_dsl doesn't really encapsulate what it does anymore - it
encodes the schema into rust.

#### What effects does this change have?

Generally renames things to be clearer: the `query_dsl` macro actually encodes a given
schema in rust, so it makes more sense for the terminology around it to be schema-centric.

- Renames `query_ds!l` -> `use_schema!`
- Changes all the docs & examples to have a mod named `schema` instead of `query_dsl`.
- Renames any other occurrences.
- Generally tries to update terminology to "schema module" or "schema"

Where possible I've left aliases for the old names, marked as deprecated - so this should
mostly be a non-breaking change. 

Fixes #222