Add `schema_for_derives` as replacement for `query_module` (#235)
#### Why are we making this change?

The `query_module` attribute macro has been around for a while, but I'm not
happy with it's name.  It doesn't capture what it's actually doing in any way.

#### What effects does this change have?

Deprecates `query_module` in favour of a new `schema_for_derives` macro that
works the same, but with a clearer name.  Also renamed the parameters of the
macro, and provided a default value for the `module` parameter.

Fixes #169