Add `spread` attribute to QueryFragment (#246)
#### Why are we making this change?

GraphQL allows the structure of queries to be shared by spreading a fragment
into a selection set.  Despite using the name `QueryFragment`, cynic does not
support this currently - a `QueryFragment` always maps 1:1 with a selection
set.

#### What effects does this change have?

Adds a `spread` field attribute to the QueryFragment derive, allowing a
`QueryFragment` to nest inside another `QueryFragment` if they both share the
same target type.