Add support for field aliases (#255)
#### Why are we making this change?

PR #253 added support for a rename attribute on individual fields of a
QueryFragment.  This opened up the possibility of selecting the same GraphQL
field multiple times in a query.

GraphQL supports doing this in queries via field aliases.  But cynic doesn't
quite understand how to do this yet - it'll just output an invalid GraphQL
query that attempts to select the same field twice.

#### What effects does this change have?

This updates the rename functionality on `QueryFragment` to be smarter when a
field ends up being selected twice via renames.  It'll automatically detect
this and request an `alias` which will result in a valid query being output.
You can also add the `alias` attribute onto any field with a `rename` to
explicitly request this.

The generator will also now do the right thing when faced with a query that
contains field aliases.

Part of #155