Fix support for idents that can't be made into raw idents (#299)
#### Why are we making this change?

Turns out that raw identifiers don't cover all keywords in rust.  They're fine
for most of them, but for some you'll get the following error:

> error: `self` cannot be a raw identifier

Apparently this is because certain keywords [are used in path resolution](https://internals.rust-lang.org/t/raw-identifiers-dont-work-for-all-identifiers/9094/4)
so even if they were made into raw keywords, things with those names would not
be accessible.

This causes problems for cynic when a graphql schema has these keywords in it,
as it assumes everything is raw-able.

#### What effects does this change have?

Updates the keyword detection code to have two different ways of handling
keywords - those that can be raw will be made raw, those that can't will have
an underscore appended to their name.  This should be transparent to users -
they can name their fields whatever they want and provide a `rename = "self"` 
and it should just work.