Add a loading spinner to the generator (#301)
The generator has quite a large wasm payload so there's often a visible delay
before its first render.  Currently this is just a blank white page, so it kind
of looks broken.  This PR adds a simple spinner to the initial HTML so it's
obvious that it is still loading.

Fixes #238