Allow schema_module attribute to be a Path (#304)
#### Why are we making this change?

Currently the schema_module attribute needs to be an identifier local to the
module of the derive.  However, at least one user has wanted to set
`schema_module` to a full module path.

#### What effects does this change have?

Updates the `schema_module` attribute to be parsed as a `syn::Path`, which 
should allow this.  I've also refactored a bunch of stuff to work in terms of
`syn::Path` rather than `TypePath` as `TypePath` wasn't really providing much
benefit.  This has also simplified a bunch of code.

Fixes #219