Reimplement InlineFragments deserialize
Now corectly supports fallback and takes the `__typename` from the
underlying QueryFragment.

Had to pilfer a bunch of buffered decoding stuff from serde, which I
suspect will make decoding slower.  Although if `__typename` is the
first field in the response (as it should be) then it should mostly
skip that.