Add a github schema crate, stop feature flagging the github example (#407)
#### Why are we making this change?

The GitHub example is currently feature flagged as working on cynic used to be
unbearable when it wasn't, as it output so much code and RA struggled to work
with it.  Things have moved on a bit now: cynic is generating less code, and
perhaps RA has also improved.

#### What effects does this change have?

Gets rid of the feature flag in the github examples, and creates a new github
schema crate so we can share a bit of code between the two examples.  I think
this might also be a good idea generally for consumers that want to work with
the GitHub API (although I'm not 100% certain).

This also acts as a good test for #373 (which seems to have been fixed by #391) 
and will ensure we don't regress it in the future.